+++ tests/serial_stubs.py
